+++ packages/navitem/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.8](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.3.7...@pluralsight/ps-design-system-navitem@1.3.8) (2020-08-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-navitem
+
+
+
+
+
 ## [1.3.7](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-navitem@1.3.6...@pluralsight/ps-design-system-navitem@1.3.7) (2020-08-24)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-navitem