+++ packages/storybook-addon-chaos/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.21](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-chaos@1.0.20...@pluralsight/ps-design-system-storybook-addon-chaos@1.0.21) (2020-08-31)
+
+**Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-chaos
+
+
+
+
+
 ## [1.0.20](https://github.com/pluralsight/design-system/compare/@pluralsight/ps-design-system-storybook-addon-chaos@1.0.19...@pluralsight/ps-design-system-storybook-addon-chaos@1.0.20) (2020-08-25)
 
 **Note:** Version bump only for package @pluralsight/ps-design-system-storybook-addon-chaos